Speed is a float; warnings solved

+++ creature.cpp
@@ -6,7 +6,7 @@ Creature::Creature(): Dynamic() {
 }
 Creature::~Creature() {
 }
-Creature::Creature(string name, float px, float py, bool solidVsSolid, bool solidVsDynamic, bool friendly, int maxHp, int maxSpeed):
+Creature::Creature(string name, float px, float py, bool solidVsSolid, bool solidVsDynamic, bool friendly, int maxHp, float maxSpeed):
 	Dynamic(name, px, py, solidVsSolid, solidVsDynamic, friendly, true, maxSpeed){
 	this->hp = maxHp;
 	this->maxHp = maxHp;
